bug: final stage终止可能出现执行成功的状态 #7023
