+++ src/backend/ci/core/process/biz-engine/src/main/kotlin/com/tencent/devops/process/engine/control/command/stage/impl/StartContainerStageCmd.kt
@@ -135,21 +135,8 @@ class StartContainerStageCmd(
         var fail: BuildStatus? = null
         var cancel: BuildStatus? = null
 
-        // 查找最后一个结束状态的Stage (排除Finally）
         val stage = commandContext.stage
-        if (stage.controlOption?.finally == true) {
-            val previousStage = pipelineStageService.listStages(stage.projectId, stage.buildId)
-                .lastOrNull {
-                    it.stageId != commandContext.stage.stageId &&
-                        (it.status.isFinish() || it.status == BuildStatus.STAGE_SUCCESS || hasFailedCheck(it))
-                }
-            // #5246 前序中如果有准入准出失败的stage则直接作为前序stage并把构建状态设为红线失败
-            commandContext.previousStageStatus = if (hasFailedCheck(previousStage)) {
-                BuildStatus.QUALITY_CHECK_FAIL
-            } else {
-                previousStage?.status
-            }
-        }
+
         // 同一Stage下的多个Container是并行
         commandContext.containers.forEach { container ->
             val jobCount = container.controlOption?.matrixControlOption?.totalCount ?: 1 // MatrixGroup存在裂变计算
